调整CreateChildren逻辑, 调整代码中对assert的使用，防止由于外部xml导致的条件不满足造成assert,以便souieditor正常使用。

+++ controls.extend/ScintillaWnd.cpp
@@ -67,7 +67,7 @@ void CScintillaWnd::UpdateLineNumberWidth(void)
 	int  iLineMarginWidthFit;  
 
 	sprintf((tchLines),(" %i "),    
-		SendMessage(SCI_GETLINECOUNT,0,0)); 
+		(int)SendMessage(SCI_GETLINECOUNT,0,0)); 
 
 	iLineMarginWidthNow = SendMessage(    
 		SCI_GETMARGINWIDTHN,0,0);
